[FE] FIX: 회원가입 시에도 프로필이미지 업로드 변경사항 적용
